// front/src/api/models/fileMetadataDto.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * DataShare API
 * API de transfert de fichiers DataShare (US01-US10)
 * OpenAPI spec version: 1.0.0
 */

export interface FileMetadataDto {
  originalName: string;
  sizeBytes: number;
  mimeType: string;
  expiresAt: string;
  passwordProtected: boolean;
}